% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cuzick_test.R
\name{cuzick.test}
\alias{cuzick.test}
\title{Cuzick's Test for Trend}
\usage{
cuzick.test(
  rank,
  alternative = c("greater", "less", "two.sided"),
  correct.ties = TRUE,
  verbose = FALSE,
  perm = FALSE,
  rsam = 10000
)
}
\arguments{
\item{rank}{The rank output from DOOR.rank function, need to include two columns: "DOOR.rank" and "label", "label" should be the ordinal group for each individual, need to be numeric from 1,2,...n}

\item{alternative}{Which type of test, default is "greater"}

\item{correct.ties}{Whether to correct ties when ties exist, default is TRUE}

\item{verbose}{Print whether ties exist or not}

\item{perm}{Whether to conduct permutation test, default is FALSE}

\item{rsam}{Number of permutations to perform, default = 1000}
}
\value{
The results of Cuzick's Test for Trend
}
\description{
Two tests can be implemented based on permutation or based on asymptotic assumptions
}
