% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOOR_rank.R
\name{DOOR.rank}
\alias{DOOR.rank}
\title{DOOR algorithm for ranking (Based on the algorithm of Evans et al (2015))}
\usage{
DOOR.rank(df, var_order, ties.method = c("average", "random"))
}
\arguments{
\item{df}{Dataset that contains all the variables that should be included for ranking and group label, class data.frame}

\item{var_order}{Desired order of variables in dataset df}

\item{ties.method}{By default, ties.method is average}
}
\value{
The rank based on DOOR algorithm based on multivariate clinical variables
}
\description{
The higher DOOR.rank, the better clinical outcome the patient has (based on benefits and harms)
}
