% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jonckheere_Terpstra_test.R
\name{JT_test}
\alias{JT_test}
\title{Jonckheere-Terpstra Test}
\usage{
JT_test(rank, rsam = 1000)
}
\arguments{
\item{rank}{The rank output from DOOR.rank function, need to include two columns: "DOOR.rank" and "label", "label" should be the ordinal group for each individual, need to be numeric from 1,2,...n}

\item{rsam}{Number of permutations to perform, default = 1000}
}
\value{
The results of Jonckheere-Terpstra Test
}
\description{
Two tests can be implemented based on permutation or based on asymptotic assumptions
}
