% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tukey_trend_test.R
\name{Tukey_trend_test}
\alias{Tukey_trend_test}
\title{Tukey's Trend Test}
\usage{
Tukey_trend_test(rank, rsam = 1000, alpha = 0.05)
}
\arguments{
\item{rank}{The rank output from DOOR.rank function, need to include two columns: "DOOR.rank" and "Dose". Make sure to have the actual dose levels for each subject, column name is "Dose", make sure it is numeric and controls have the lowest dose level}

\item{rsam}{Number of permutations to perform, default = 1000}

\item{alpha}{Desired type I error level, defalt = 0.05}
}
\value{
The results of Tukey's Trend Test
}
\description{
Two tests can be implemented based on permutation or based on asymptotic assumptions
}
